backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
django-cors-headers